pyproject.toml
[project]
name = "kakeibo-be"
version = "0.1.0"
description = ""
authors = [
    {name = "nana-ichimaru",email = "nana.ichimaru@novel.jpn.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[all] (>=0.124.0,<0.125.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "kakeibo_be", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.8,<0.15.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]

[tool.ruff]
extend-select = [
  "E",    # pycodestyle（PEP8）
  "F",    # pyflakes（未使用の変数など）
  "I",    # isort（インポート順）
  "B",    # flake8-bugbear（バグにつながるコード）
  "UP",   # pyupgrade（新しい構文の推奨）
  "C4",   # flake8-comprehensions（内包表記の改善）
  "ANN",  # flake8-annotations（型アノテーションの有無）
  "SIM",  # flake8-simplify（冗長な構文の簡略化）
  "FAST002",
  "FAST003",  # fast-api-unused-path-parameter（FastAPIにおいてpath parameterのシグネチャ宣言を強制）
]

ignore = [
  "E501",    # 行が長すぎる
  "B008"
]

line-length = 100  # 行の最大長
target-version = "py314"  # 対象Pythonバージョン
show-fixes = true         # VSCode上でfixできる項目を表示

[tool.ruff.isort]
known-first-party = ["kakeibo_be"] # 自作モジュールの名前
combine-as-imports = true
force-single-line = false
lines-between-types = 1

[tool.ruff.flake8-type-checking]
strict = true  # 型チェック目的のインポートはすべて遅延させるように警告
